refactor WDT helper classes
