+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/wdt/WdtOperation.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2019, Oracle Corporation and/or its affiliates.  All rights reserved.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.util;
+package com.oracle.weblogic.imagetool.wdt;
 
 /**
  * Command types for WebLogic Deploy Tooling.
