+++ imagetool/src/main/resources/ImageTool.properties
